dom/base/test/test_bug390219.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=390219
-->
<head>
  <meta http-equiv="content-type" content="text/html; charset=utf-8">
  <title>Test for Bug 390219</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 390219 **/

SimpleTest.waitForExplicitFinish();
xhr = new XMLHttpRequest();
xhr.open("GET", "nonexistent_url", true);
xhr.send(null);
xhr.abort();
xhr.open("GET", ".", true);
xhr.onreadystatechange = function() {
  if (xhr.readyState == 4) {
    is(xhr.status, 200, "wrong status");
    SimpleTest.finish();
  }
}
xhr.send(null);

</script>
</pre>
</body>
</html>
